fix(export): render Mermaid diagrams in image exports (#728)
